requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.4
commonmark==0.9.1
crashtest==0.3.1
distlib==0.3.4
dynaconf==3.1.8
et-xmlfile==1.1.0
filelock==3.6.0
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
keyring==23.5.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==1.0.3
mysql-connector-python==8.0.29
numpy==1.22.4
openpyxl==3.0.10
packaging==20.9
pandas==1.4.2
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.1.13
poetry-core==1.0.8
protobuf==3.20.1
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pylev==1.4.0
PyMySQL==1.0.2
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.1
shellingham==1.4.0
six==1.16.0
SQLAlchemy==1.4.36
tomli==2.0.1
tomlkit==0.10.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
zipp==3.8.0
